APM 对数通过

+++ models/src/main/scala/com/pharbers/models/entity/apmreport.scala
@@ -5,15 +5,15 @@ import com.pharbers.macros.common.connecting.ToStringMacro
 
 @ToStringMacro
 class apmreport extends commonEntity {
-    var potential: Long = 0L
+    var potential: Double = 0.0
     var potential_growth: Double = 0.0
     var potential_contri: Double = 0.0
-    var sales: Long = 0L
+    var sales: Double = 0.0
     var sales_growth: Double = 0.0
     var sales_contri: Double = 0.0
     var contri_index: Double = 0.0
     var share: Double = 0.0
     var share_change: Double = 0.0
-    var company_target: Long = 0L
+    var company_target: Double = 0.0
     var achieve_rate: Double = 0.0
 }
